refactor(edge): isolate request preparation flow from forwarding
